Update infiblocks.py
Introduce player extension packet, which is already a thing in BetterSpades. This sends an update of some player specific numbers, allowing to modify the number of blocks and grenades to what they are. 

This fixes a bug that made scrolling through tools with 0 grenades + refilled blocks kinda weird, to all clients that support this packet or will eventually support it.